examples/discordpy_example/requirements.txt
discord.py
topggpy